requirements.txt
pytesseract
Pillow
numpy
pyautogui
pydirectinput
